Refactor Fabric API extension
