+++ src/main/java/net/fabricmc/loom/configuration/fabricapi/FabricApiDataGeneration.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of fabric-loom, licensed under the MIT License (MIT).
  *
- * Copyright (c) 2020-2023 FabricMC
+ * Copyright (c) 2024 FabricMC
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,86 +22,38 @@
  * SOFTWARE.
  */
 
-package net.fabricmc.loom.configuration;
+package net.fabricmc.loom.configuration.fabricapi;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
 import javax.inject.Inject;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
 
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.ConfigurationContainer;
-import org.gradle.api.artifacts.Dependency;
-import org.gradle.api.file.RegularFileProperty;
 import org.gradle.api.plugins.JavaPlugin;
-import org.gradle.api.provider.Property;
 import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.SourceSetContainer;
 import org.gradle.api.tasks.TaskContainer;
 import org.gradle.jvm.tasks.Jar;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.NodeList;
 
 import net.fabricmc.loom.LoomGradleExtension;
+import net.fabricmc.loom.api.fabricapi.DataGenerationSettings;
 import net.fabricmc.loom.configuration.providers.minecraft.MinecraftSourceSets;
-import net.fabricmc.loom.util.download.DownloadException;
 import net.fabricmc.loom.util.fmj.FabricModJson;
 import net.fabricmc.loom.util.fmj.FabricModJsonFactory;
 import net.fabricmc.loom.util.gradle.SourceSetHelper;
 
-public abstract class FabricApiExtension {
-	@Inject
-	public abstract Project getProject();
-
+abstract class FabricApiDataGeneration {
 	private static final String DATAGEN_SOURCESET_NAME = "datagen";
 
-	private static final HashMap<String, Map<String, String>> moduleVersionCache = new HashMap<>();
-	private static final HashMap<String, Map<String, String>> deprecatedModuleVersionCache = new HashMap<>();
+	@Inject
+	protected abstract Project getProject();
 
-	public Dependency module(String moduleName, String fabricApiVersion) {
-		return getProject().getDependencies()
-				.create(getDependencyNotation(moduleName, fabricApiVersion));
-	}
-
-	public String moduleVersion(String moduleName, String fabricApiVersion) {
-		String moduleVersion = moduleVersionCache
-				.computeIfAbsent(fabricApiVersion, this::getApiModuleVersions)
-				.get(moduleName);
-
-		if (moduleVersion == null) {
-			moduleVersion = deprecatedModuleVersionCache
-					.computeIfAbsent(fabricApiVersion, this::getDeprecatedApiModuleVersions)
-					.get(moduleName);
-		}
-
-		if (moduleVersion == null) {
-			throw new RuntimeException("Failed to find module version for module: " + moduleName);
-		}
-
-		return moduleVersion;
-	}
-
-	/**
-	 * Configure data generation with the default options.
-	 */
-	public void configureDataGeneration() {
-		configureDataGeneration(dataGenerationSettings -> { });
-	}
-
-	/**
-	 * Configure data generation with custom options.
-	 */
-	public void configureDataGeneration(Action<DataGenerationSettings> action) {
+	void configureDataGeneration(Action<DataGenerationSettings> action) {
 		final LoomGradleExtension extension = LoomGradleExtension.get(getProject());
 		final TaskContainer taskContainer = getProject().getTasks();
 
@@ -195,127 +147,6 @@ public abstract class FabricApiExtension {
 			getProject().getTasks().named("runDatagen", task -> {
 				task.getOutputs().dir(outputDirectory);
 			});
-		}
-	}
-
-	public interface DataGenerationSettings {
-		/**
-		 * Contains the output directory where generated data files will be stored.
-		 */
-		RegularFileProperty getOutputDirectory();
-
-		/**
-		 * Contains a boolean indicating whether a run configuration should be created for the data generation process.
-		 */
-		Property<Boolean> getCreateRunConfiguration();
-
-		/**
-		 * Contains a boolean property indicating whether a new source set should be created for the data generation process.
-		 */
-		Property<Boolean> getCreateSourceSet();
-
-		/**
-		 * Contains a string property representing the mod ID associated with the data generation process.
-		 *
-		 * <p>This must be set when {@link #getCreateRunConfiguration()} is set.
-		 */
-		Property<String> getModId();
-
-		/**
-		 * Contains a boolean property indicating whether strict validation is enabled.
-		 */
-		Property<Boolean> getStrictValidation();
-
-		/**
-		 * Contains a boolean property indicating whether the generated resources will be automatically added to the main sourceset.
-		 */
-		Property<Boolean> getAddToResources();
-
-		/**
-		 * Contains a boolean property indicating whether data generation will be compiled and ran with the client.
-		 */
-		Property<Boolean> getClient();
-	}
-
-	private String getDependencyNotation(String moduleName, String fabricApiVersion) {
-		return String.format("net.fabricmc.fabric-api:%s:%s", moduleName, moduleVersion(moduleName, fabricApiVersion));
-	}
-
-	private Map<String, String> getApiModuleVersions(String fabricApiVersion) {
-		try {
-			return populateModuleVersionMap(getApiMavenPom(fabricApiVersion));
-		} catch (PomNotFoundException e) {
-			throw new RuntimeException("Could not find fabric-api version: " + fabricApiVersion);
-		}
-	}
-
-	private Map<String, String> getDeprecatedApiModuleVersions(String fabricApiVersion) {
-		try {
-			return populateModuleVersionMap(getDeprecatedApiMavenPom(fabricApiVersion));
-		} catch (PomNotFoundException e) {
-			// Not all fabric-api versions have deprecated modules, return an empty map to cache this fact.
-			return Collections.emptyMap();
-		}
-	}
-
-	private Map<String, String> populateModuleVersionMap(File pomFile) {
-		try {
-			DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
-			DocumentBuilder docBuilder = docFactory.newDocumentBuilder();
-			Document pom = docBuilder.parse(pomFile);
-
-			Map<String, String> versionMap = new HashMap<>();
-
-			NodeList dependencies = ((Element) pom.getElementsByTagName("dependencies").item(0)).getElementsByTagName("dependency");
-
-			for (int i = 0; i < dependencies.getLength(); i++) {
-				Element dep = (Element) dependencies.item(i);
-				Element artifact = (Element) dep.getElementsByTagName("artifactId").item(0);
-				Element version = (Element) dep.getElementsByTagName("version").item(0);
-
-				if (artifact == null || version == null) {
-					throw new RuntimeException("Failed to find artifact or version");
-				}
-
-				versionMap.put(artifact.getTextContent(), version.getTextContent());
-			}
-
-			return versionMap;
-		} catch (Exception e) {
-			throw new RuntimeException("Failed to parse " + pomFile.getName(), e);
-		}
-	}
-
-	private File getApiMavenPom(String fabricApiVersion) throws PomNotFoundException {
-		return getPom("fabric-api", fabricApiVersion);
-	}
-
-	private File getDeprecatedApiMavenPom(String fabricApiVersion) throws PomNotFoundException {
-		return getPom("fabric-api-deprecated", fabricApiVersion);
-	}
-
-	private File getPom(String name, String version) throws PomNotFoundException {
-		final LoomGradleExtension extension = LoomGradleExtension.get(getProject());
-		final var mavenPom = new File(extension.getFiles().getUserCache(), "fabric-api/%s-%s.pom".formatted(name, version));
-
-		try {
-			extension.download(String.format("https://maven.fabricmc.net/net/fabricmc/fabric-api/%2$s/%1$s/%2$s-%1$s.pom", version, name))
-					.defaultCache()
-					.downloadPath(mavenPom.toPath());
-		} catch (DownloadException e) {
-			if (e.getStatusCode() == 404) {
-				throw new PomNotFoundException(e);
-			}
-
-			throw new UncheckedIOException("Failed to download maven info to " + mavenPom.getName(), e);
-		}
-
-		return mavenPom;
-	}
-
-	private static class PomNotFoundException extends Exception {
-		PomNotFoundException(Throwable cause) {
-			super(cause);
 		}
 	}
